Use byte array layout for initialising GC heap
When initially allocating memory for the main GC heap, we don't know (or
care) about the memory layout of the objects that will be stored there.
That is the role of the allocator when placing individual objects in the
heap.

For this reason, the most straightforward approach is to allocate a
chunk of memory as a byte array and leave the interpretation of the
individual blocks up to the collector at runtime.